export-data: Convert newlines so that they don't break the csv
If we have a value such as an address which has lots of new lines we
need to make sure theses aren't interpreted as new rows when loading
this as a csv. Also filter the data to make sure that when displaying it
in the admin page it isn't going to cause any xss.

Also fix a few comments